fix some pylint issues
